refactor(coding-agent): extract agent-session permissions module
Move the ACP permission-gate pure helpers (PERMISSION_REQUIRED_TOOLS,
PERMISSION_OPTIONS, PERMISSION_OPTIONS_BY_ID, getStringProperty,
getEditDestructiveIntent, getPermissionIntent, extractPermissionLocations) into
sibling agent-session-permissions.ts. These carry no session state; AgentSession
imports back the six symbols it references. Not part of the public surface, so no
re-export and the runtime export surface is byte-identical.

Op: compress
